[refactor] convert all routes to typescript
